style: add layout and change folders
