+++ src/app/(private-pages)/agenda/page.tsx
@@ -1,0 +1,12 @@
+import EventList from '../../../components/eventList';
+
+export default function Agenda() {
+  return (
+    <>
+      <h1 className="text-3xl font-semibold p-4">Agenda</h1>
+
+      <EventList />
+      <EventList />
+    </>
+  );
+}
