+++ src/app/(private-pages)/calendario/page.tsx
@@ -1,0 +1,11 @@
+import CalendarHeader from '../../../components/MonthCalendar/CalendarHeader';
+import MonthCalendar from '../../../components/MonthCalendar/MonthCalendar';
+
+export default function Calendar() {
+  return (
+    <>
+      <CalendarHeader />
+      <MonthCalendar />
+    </>
+  );
+}
