+++ src/app/(private-pages)/cursos/(cursos-pages)/Hero.tsx
@@ -14,7 +14,7 @@ export default function Hero({ title }: HeroProps) {
 
   return (
     <motion.header
-      className="bg-gradient-to-b from-green-800 via-green-600 to-green-800 text-white h-52 flex items-center justify-between px-6"
+      className="flex items-center justify-between"
       initial={{ opacity: 0, scale: 0.5 }}
       animate={{ opacity: 1, scale: 1 }}
       transition={{
@@ -23,9 +23,9 @@ export default function Hero({ title }: HeroProps) {
         ease: [0, 0.71, 0.2, 1.01],
       }}
     >
-      <h1 className="text-5xl font-bold">Fap - {title}</h1>
+      <h1 className="text-3xl font-bold ">Fap - {title}</h1>
       <button
-        className="hover:bg-green-800 focus:bg-green-700 rounded-full w- flex justify-center items-center p-1"
+        className="hover:bg-gray-300 focus:bg-gray-400 rounded-full w- flex justify-center items-center p-1"
         onClick={() => router.push('/cursos')}
       >
         <IoIosArrowBack className="text-4xl" />
