+++ src/app/(private-pages)/cursos/(cursos-pages)/Syllabus.tsx
