+++ src/app/(private-pages)/cursos/(cursos-pages)/desenvolvedor-front-end/page.tsx
