+++ src/app/(private-pages)/cursos/CoursesList.tsx
