+++ src/app/(private-pages)/cursos/page.tsx
