--- src/app/agenda/page.tsx
@@ -1,15 +1,0 @@
-/* eslint-disable prettier/prettier */
-
-import EventList from "../components/eventList";
-
-export default function Agenda(){
-    return( <>
-
-    
-    <h1 className="text-3xl font-semibold p-4">Agenda</h1>
-
-    <EventList />
-    <EventList />
-    
-    </>)
-}
