--- src/app/calendario/page.tsx
@@ -1,11 +1,0 @@
-import CalendarHeader from '../components/MonthCalendar/CalendarHeader';
-import MonthCalendar from '../components/MonthCalendar/MonthCalendar';
-
-export default function Calendar() {
-  return (
-    <>
-      <CalendarHeader />
-      <MonthCalendar />
-    </>
-  );
-}
