--- src/app/cursos/(cursos-pages)/layout.tsx
@@ -1,7 +1,0 @@
-export default function CoursesLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return <section>{children}</section>;
-}
