+++ src/app/layout.tsx
@@ -9,7 +9,7 @@ import CalendarContextWrapper from '@/contexts/CalendarContextWrapper';
 const inter = Inter({ subsets: ['latin'] });
 
 export const metadata: Metadata = {
-  title: 'Create Next App',
+  title: 'Timesync',
   description: 'Generated by create next app',
 };
